macos-app/requirements.txt
pynput==1.7.6
Pillow==10.1.0
